[新增] 怪物血量隨關卡與波數增長

+++ Assets/Scripts/Game/Enemy/EnemyCombinedJob.cs
@@ -22,7 +22,7 @@ public struct EnemyJobData
     public int InstanceID;
     public EnemyMoveType MoveType;
 
-    public float CurrentHp;
+    public int CurrentHp;
     public float MoveSpeed;
     public float AttackRange;
     public float Radius;
